requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pytest>=7.4.0
